Remove auto-closing and lazyReader
Both of these were a little too magical for my liking. "Closing" a part
is the responsibility of whoever opened it; Stream shouldn't be doing
that for you. Definitely will be adding a convenience function for doing
a similar thing. But the core should be svelte.

This commit also renames things. Imo makes it a nicer api to use.